requirements.txt
pandas
pyarrow
pymongo
django 
djangorestframework 
djongo

